Extract cookie operation
